Fix layout and pages
