+++ csci32-nextjs-stimpson/src/app/.gitattributes
