+++ csci32-nextjs-stimpson/src/app/README.md
