+++ csci32-nextjs-stimpson/src/app/about/page.tsx
@@ -1,6 +1,6 @@
-import Head from 'next/head';
-import React from 'react'; // Importing React to use React.Fragment
-import Image from 'next/image'; // Importing Image component from Next.js
+import Head from 'next/head'
+import React from 'react' // Importing React to use React.Fragment
+import Image from 'next/image' // Importing Image component from Next.js
 
 export default function Portfolio() {
   return (
@@ -13,34 +13,14 @@ export default function Portfolio() {
         <link href="/dist/output.css" rel="stylesheet" />
       </Head>
 
-      {/* Header */}
-      <header className="bg-white shadow">
-        <div className="container mx-auto p-6">
-          <h1 className="text-3xl font-bold">Painter's Name</h1>
-          <nav>
-            <ul className="flex space-x-4">
-              <li>
-                <a href="/" className="text-blue-500">Home</a>
-              </li>
-              <li>
-                <a href="/about.html" className="text-blue-500">About</a>
-              </li>
-              <li>
-                <a href="/contact.html" className="text-blue-500">Contact</a>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </header>
-
       {/* Main Content */}
       <main className="container mx-auto py-12">
         <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
           <div className="flex flex-col justify-center">
             <h2 className="text-4xl font-semibold mb-4">Transforming Spaces with Color</h2>
             <p className="text-lg">
-              Welcome to the portfolio of a passionate painter. Check out my latest works and get in touch for an estimate
-              on your next project.
+              Welcome to the portfolio of a passionate painter. Check out my latest works and get in touch for an
+              estimate on your next project.
             </p>
             <a href="/estimate.html" className="mt-6 inline-block bg-blue-500 text-white py-2 px-4 rounded">
               Get an Estimate
@@ -63,5 +43,5 @@ export default function Portfolio() {
         <p>© 2024 Painter's Name. All rights reserved.</p>
       </footer>
     </React.Fragment>
-  );
+  )
 }
