+++ csci32-nextjs-stimpson/src/app/favicon.ico
