+++ csci32-nextjs-stimpson/src/app/globals.css
