+++ csci32-nextjs-stimpson/src/app/hiddentab.tsx
