--- csci32-nextjs-stimpson/src/app/week1-page/page.tsx/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "semi": false,
-  "singleQuote": true,
-  "tabWidth": 2,
-  "useTabs": false
-}
