--- csci32-nextjs-stimpson/src/app/week1-page/page.tsx/next.config.js
@@ -1,4 +1,0 @@
-// next.config.js
-module.exports = {
-  reactStrictMode: true,
-};
