--- csci32-nextjs-stimpson/src/app/week1-page/page.tsx/package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "prettier": "^3.3.3"
-  }
-}
