--- csci32-nextjs-stimpson/src/app/week1-page/page.tsx/postcss.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-}
