--- csci32-nextjs-stimpson/src/app/week1-page/page.tsx/tailwind.config.js
@@ -1,11 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ['./app/**/*.{js,ts,jsx,tsx}',
-    './pages/**/*.{js,ts,jsx,tsx}',
-    './components/**/*.{js,ts,jsx,tsx}',],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
-
